### 2025-05-25 #8270 @editorialbot
[PRE REVIEW]: ``OrbDot``: A Python package for studying the secular evolution of exoplanet orbits
**Submitting author:** @simonehagey (<a href="http://orcid.org/0000-0001-8072-0590">Simone Hagey</a>)
**Repository:** https://github.com/simonehagey/orbdot
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0rc1
