### 2025-05-26 fec5805
added JOSS badge to README
