### 2025-05-26 5169ea6
added JOSS badge to README
