+++ README.rst
@@ -56,9 +56,10 @@ Feedback, bug reports, feature requests, and code contributions are welcome! Ple
 
 -----
 
+|JOSS Status|
+
 TrES-1 b Case Study
 ===================
-|JOSS Status|
 In addition to the documentation, there is a complementary `TrES-1 b case study <LINK>`__ that showcases the full functionality of OrbDot’s model-fitting and interpretive tools, while providing a deeper dive into the theoretical foundations behind the package. This paper consolidates key equations in one place, making it a valuable reference for researchers looking to apply OrbDot to their own work.
 
 
