### 2025-06-23 #8270 @simonehagey
Hi @warrickball, the AAS paper associated with OrbDot has just been accepted for publication. I was wondering if you have any update on the availability of a suitable JOSS editor? Thanks!
