### 2025-06-24 #8270 @warrickball
Hi @simonehagey, sorry for the slight delay. I'll act as the editor on this and start looking for reviewers in the next hour or so. With a bit of luck it'll only take a few days to find reviewers but, as you say, I'll start the review early if necessary, since we can still add reviewers after the review issue opens. It's just slightly neater for our database if we have them in place beforehand.
