### 2025-07-07 #8550 @simonehagey
Thank you @soichiro-hattori and @j-faria for agreeing to review!

@warrickball is it okay if I send this pre-assigned DOI (10.21105/joss.08550) to the AAS editor so that the companion paper can be sent to their production team?
