### 2025-07-15 #8550 @warrickball
Hi @soichiro-hattori and @j-faria, I just wanted to check in to see if you were progressing with your reviews. Note that at JOSS we encourage iterative reviews, where you raises issues as you encounter them, rather than monolithic reviews like more mainstream journals. You're welcome to bring things up here or in the [software repo's issues](https://github.com/simonehagey/orbdot/issues), where you can tag this review by hyperlinking back to it.

If you have any questions for me, don't hesitate to ask!
