### 2025-07-16 simonehagey/orbdot#7 @soichiro-hattori
Potential leftover sentence in WASP-12 b example docs
The following sentence seems to be out of place. Perhaps it should be removed?

https://github.com/simonehagey/orbdot/blob/23dd7577e280131cef4ea7dcbf68223e1adfdb2c/docs/source/example_wasp-12.rst?plain=1#L173

Ref: https://github.com/openjournals/joss-reviews/issues/8550
