### 2025-07-17 f802bdb
Updated AAS paper DOI
