+++ joss/paper.bib
@@ -56,3 +56,11 @@
        adsurl = {https://ui.adsabs.harvard.edu/abs/2021ascl.soft03022B},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+@ARTICLE{Hagey:2025,
+       author = {{Hagey}, Simone R. and {Edwards}, Billy and {Tsiaras}, Angelos and and {Boley}, Aaron and others},
+        title = "{TrES-1 b: A Case Study in Detecting Secular Evolution of Exoplanet Orbits}",
+      journal = {The Astronomical Journal},
+         year = 2025,
+          doi = {10.3847/1538-3881/aded15},
+}
