### 2025-07-24 #8550 @soichiro-hattori
@simonehagey: I have two questions right now!
1. Are the three priors currently supported (gaussian, log-uniform, uniform) enough to cover most realistic use cases? Or are there plans for adding more?
2. Is there any documentation about the performance of the nested sampling or its dependence on the number of parameters you might be fitting for? I don't have too much experience with nested sampling but my impression was that the sampling efficiency degrades as a function of the number of model parameters.

Thank you!
