### 2025-07-30 #8550 @soichiro-hattori
@simonehagey: Thank you for your response!

I think the documentation is already great and comprehensive, but maybe if you have the time then mentioning these things somewhere in the docs, maybe like a troubleshooting or FAQ page, might be helpful for a user who's less acquainted with model fitting and nested sampling (perhaps you could link the review article)?

Could you perhaps link the Github repo on the main landing page in the docs?
