### 2025-08-12 simonehagey/orbdot#8 @j-faria
Plotting error running HAT-P-4 example
When running the [HAT-P-4 example](https://orbdot.readthedocs.io/en/latest/example_rv_trends.html#hat-p-4-b) as described in the documentation, I got the following error when doing the plots:

```py
Initializing HAT-P-4b instance...
[...]
----------------------------------------------------------------------------------------------------
Generating RV plot...
----------------------------------------------------------------------------------------------------
2 RV source(s) detected: ['Bonomo et al. 2017' 'Knutson et al. 2014'] and assigned tag(s): ['Bon', 'Knu']

Traceback (most recent call last):
  File "example_hatp4.py", line 31, in <module>
    fit_circular = hatp4.run_rv_fit(["t0", "P0", "K", "v0", "jit"], file_suffix="_circular")
                   ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "orbdot\radial_velocity.py", line 216, in run_rv_fit
    res = self.run_rv_constant(free_params, file_suffix, make_plot)
          ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "orbdot\radial_velocity.py", line 365, in run_rv_constant
    pl.make_rv_plots(
  File "orbdot\tools\plots.py", line 1017, in make_rv_plots
    s_orb, s_tdp, s_rv = read_random_samples(samples_file)
                         ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "orbdot\tools\plots.py", line 2156, in read_random_samples
    rdv = [float(row[11])]
                 ~~~^^^^
IndexError: list index out of range
```

(opened as part of the JOSS review openjournals/joss-reviews#8550)
