### 2025-08-12 4e886e0
Add tests for RV model fits
