### 2025-08-12 bfd4e58
fixed Windows newline bug when reading/writing files
