+++ .github/workflows/test.yaml
@@ -14,6 +14,7 @@ jobs:
       matrix:
         os: [ubuntu-latest, macos-latest, windows-latest]
         python-version: ["3.9", "3.10", "3.13"]
+        fail-fast: false
 
     runs-on: ${{ matrix.os }}
     steps:
