### 2025-08-13 e65aea3
added direct download link for /examples dir
