### 2025-08-13 7c6217e
added a github link to the front page
