### 2025-08-14 ee149e6
added more nested sampling info to docs
