+++ docs/source/references.bib
@@ -538,3 +538,16 @@ archivePrefix = {ascl},
         pages = {10},
           doi = {10.21105/astro.1306.2144},
 }
+
+@article{Buchner2023,
+       author = {Johannes Buchner},
+       title = {{Nested sampling methods}},
+       volume = {17},
+       journal = {Statistics Surveys},
+       number = {none},
+       publisher = {Amer. Statist. Assoc., the Bernoulli Soc., the Inst. Math. Statist., and the Statist. Soc. Canada},
+       pages = {169 -- 215},
+       year = {2023},
+       doi = {10.1214/23-SS144},
+       URL = {https://doi.org/10.1214/23-SS144}
+}
