### 2025-08-14 #8550 @simonehagey
> I think the documentation is already great and comprehensive, but maybe if you have the time then mentioning these things somewhere in the docs, maybe like a troubleshooting or FAQ page, might be helpful for a user who's less acquainted with model fitting and nested sampling (perhaps you could link the review article)?

@soichiro-hattori Thank you, and apologies for the delay in getting back to you!

I agree that the documentation could benefit from more background on nested sampling. I decided that an appropriate place to add more info and suggest resources was in the "Getting Started" page, where it describes the settings for model fitting (https://orbdot.readthedocs.io/en/latest/getting_started.html#the-settings-file). Let me know if you think another location would be more appropriate or if additional detail would be helpful.

> Could you perhaps link the Github repo on the main landing page in the docs?

Done, thanks for the suggestion!
