### 2025-08-14 #8550 @simonehagey
> I can certainly tick it already, but I noticed that the example files (info, settings, data) are not easily accessible from a PyPI installation of `orbdot` and they need to be downloaded from the repository (sorry if I missed something!). If possible, in order to lower the barrier of entry, I would suggest either
>
> 1. to include the example files with the PyPI installation, and make them accessible directly through the package
> 2. to include the files verbatim in the documentation pages (e.g. [here](https://orbdot.readthedocs.io/en/latest/example_wasp-12.html)) so the user can copy them

@j-faria thank you for pointing this out! I had originally tried including the example files in the installation, but handling the relative file path names was really frustrating and it only works if you clone the repository.

To make it more accessible, I decided to create a direct link for downloading a .zip file of the “examples/“ directory, which I included in three locations in the docs:

1. At the top of the orbital decay example (https://orbdot.readthedocs.io/en/latest/example_wasp-12.html)
2. At the top of the RV fit example (https://orbdot.readthedocs.io/en/latest/example_rv_trends.html)
3. On the Installation page in the “Test the Installation” section (https://orbdot.readthedocs.io/en/latest/installation.html#test-the-installation)

Do you think this is an appropriate solution?

> Do I understand correctly that there is currently only one automated test in https://github.com/simonehagey/orbdot/blob/main/tests/test_decay.py ?
>
> Could the provided examples be added as automated tests as well?

@j-faria I addressed this suggestion in tandem with fixing the bug that you found, which really highlighted how important it is to have more test cases :) The issue ended up being specific to Windows, so in addition to fixing the bug I added Windows to the list of operating systems for the PyTest workflows.

I also added unit tests that run RV model fits for HAT-P-4 b and HAT-P-22 b, as you suggested. All of the unit tests are passing now and cover the MacOS/Ubuntu/Windows operating systems and Python versions 3.9/3.10/3.11.
