### 2025-08-14 #8550 @simonehagey
@soichiro-hattori @j-faria thank you again for volunteering your time to this review, I know it was a lot of work to read through everything and I really appreciate your feedback!

In the checklists, I noticed that both of you left the “State of the field” box unchecked in the Software Paper section. Please let me know if you would like me to make any changes to the paper.

@warrickball I’m preparing to submit my comments on the proofs for the AAS paper and realized that, in addition to the JOSS DOI, they also require a volume and issue number. Based on other JOSS papers, I determined the following:

year = {2025},
volume = {10},
number = {112},
pages = {8550},

The only point I’m unsure about is whether to list the issue number as 112 (August) or 113 (September). Given the progress we’ve made in this review so far, do you think it’s reasonable to use 112 in the AAS proofs? I can also ask them to hold off a bit longer, no pressure, I’m just unfamiliar with how best to handle this type of situation. Thanks!
