### 2025-08-15 #8550 @j-faria
Hi @simonehagey

Thank you, it's definitely much easier to run the examples now!
I apologize for not providing a whole lot of information on the plotting issue. It didn't cross my mind it could be specific to Windows. But I'm glad you were able to solve it.

> In the checklists, I noticed that both of you left the “State of the field” box unchecked in the Software Paper section. Please let me know if you would like me to make any changes to the paper.

I personally hadn't check this box because it specifically mentions "how this software compares to other commonly-used packages", and I couldn't find information about this in the paper. At the same time, I suspect there are few other packages with the same features as `orbdot` (there could be a small overlap with what we have in [`kima`](https://github.com/kima-org/kima), related to apsidal precession, as in [Baycroft et al. 2023](https://ui.adsabs.harvard.edu/abs/2023MNRAS.521.1871B/abstract)?). In any case, I'm very happy to accept the paper as is.
