### 2025-08-15 #8550 @j-faria
Pinging @warrickball - I'm happy to accept the paper as is.
Thank you @simonehagey for your responses and changes to the code. I'm looking forward to using `orbdot` in the future!
