### 2025-08-15 #8550 @warrickball
Hi @simonehagey, regarding the volume number, I'm about to go on holiday until 25 August so I think August publication is unlikely and issue number 113 makes more sense. There'll be at least a few [post-review items](https://joss.readthedocs.io/en/latest/editing.html#post-review) to work through. We still need @soichiro-hattori to comment on his review.

Regarding @j-faria's comment about other software, even if there isn't any similar software, it's worth stating how the closest-related ones still leave a gap that OrbDot fills. E.g. commenting on the overlap with apsidal procession in Kima is a good start. I suspect [Exostriker](https://exostriker-manual.readthedocs.io/) overlaps a bit too, but I'm no expert.
