### 2025-08-19 #8550 @soichiro-hattori
> I agree that the documentation could benefit from more background on nested sampling. I decided that an appropriate place to add more info and suggest resources was in the "Getting Started" page, where it describes the settings for model fitting (https://orbdot.readthedocs.io/en/latest/getting_started.html#the-settings-file). Let me know if you think another location would be more appropriate or if additional detail would be helpful.

I think that works!

> In the checklists, I noticed that both of you left the “State of the field” box unchecked in the Software Paper section. Please let me know if you would like me to make any changes to the paper.

I also hadn't checked that box for the reason @j-faria mentioned, but am happy to since it seems like you will be incorpoting @warrickball's suggestion!

@warrickball - I'm also happy to accept the project/paper!
