### 2025-09-02 0818726
updates to JOSS paper
