+++ joss/paper.bib
@@ -57,6 +57,159 @@
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@INPROCEEDINGS{Barker:2024,
+       author = {{Barker}, Malia and {Jackson}, Brian and {Huchmala}, Rachel and {Adams}, Elisabeth and {Kirk}, Adrienne},
+        title = "{Susie Transiting Exoplanet Ephemeris Package}",
+    booktitle = {56th Annual Meeting of the Division for Planetary Sciences},
+         year = 2024,
+       series = {AAS Division for Planetary Sciences Meeting Abstracts},
+       volume = {56},
+        month = oct,
+          eid = {402.02},
+        pages = {402.02},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2024DPS....5640202B},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@MISC{susie,
+      author = {{Barker}, Malia and {Kirk}, Adrienne},
+       title = {The Susie Python Package},
+        year = {2025},
+   publisher = {GitHub},
+     journal = {GitHub repository},
+howpublished = {\url{https://github.com/BoiseStatePlanetary/susie.git}},
+}
+
+@ARTICLE{Wang:2024,
+       author = {{Wang}, Wenqin and {Zhang}, Zixin and {Chen}, Zhangliang and {Wang}, Yonghao and {Yu}, Cong and {Ma}, Bo},
+        title = "{Long-term Variations in the Orbital Period of Hot Jupiters from Transit-timing Analysis Using TESS Survey Data}",
+      journal = {The Astrophysical Journal Supplement Series},
+         year = 2024,
+        month = jan,
+       volume = {270},
+       number = {1},
+          eid = {14},
+        pages = {14},
+          doi = {10.3847/1538-4365/ad0847},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2024ApJS..270...14W},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Deck:2014,
+       author = {{Deck}, Katherine M. and {Agol}, Eric and {Holman}, Matthew J. and {Nesvorn{\'y}}, David},
+        title = "{TTVFast: An Efficient and Accurate Code for Transit Timing Inversion Problems}",
+      journal = {The Astrophysical Journal},
+         year = 2014,
+        month = jun,
+       volume = {787},
+       number = {2},
+          eid = {132},
+        pages = {132},
+          doi = {10.1088/0004-637X/787/2/132},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2014ApJ...787..132D},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Espinoza:2019,
+       author = {{Espinoza}, N{\'e}stor and {Kossakowski}, Diana and {Brahm}, Rafael},
+        title = "{juliet: a versatile modelling tool for transiting and non-transiting exoplanetary systems}",
+      journal = {Monthly Notices of the Royal Astronomical Society},
+         year = 2019,
+        month = dec,
+       volume = {490},
+       number = {2},
+        pages = {2262-2283},
+          doi = {10.1093/mnras/stz2688},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2019MNRAS.490.2262E},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{ForemanMackey:2021,
[…]
