### 2025-09-02 a97cb58
added TrES-1 b paper arXiv link
