### 2025-09-02 f908116
JOSS paper edits
