+++ joss/paper.md
@@ -33,9 +33,9 @@ Gradual changes in exoplanet orbits, known as secular variations, can be detecte
 
 The necessity to search over many orbital epochs coupled with an ever-growing archive of exoplanet observations creates a need for fast and flexible open-source software that can reliably detect gradual changes in exoplanet orbits. `OrbDot` addresses this need by offering robust tools for fitting secular evolution models to exoplanet transit and eclipse mid-times, transit durations, and radial velocity data.
 
-A key advantage of `OrbDot` is its ability to fit multiple data types simultaneously, which can help to break parameter degeneracies. It also excels at assisting in result interpretation by generating reports on model comparisons and assessments of various physical effects in the context of the models and their corresponding theory. For example, analysis reports could determine key parameters for assessing tidal energy dissipation, apsidal precession mechanisms, variations due to systemic proper motion, and the dynamical effects of non-resonant companion objects, depending on the applied models.
+A key advantage of `OrbDot` is its ability to fit multiple types of data simultaneously, which can help to break parameter degeneracies. It also excels at assisting in result interpretation by generating reports on model comparisons and assessments of various physical effects in the context of the models and their corresponding theory. For example, analysis reports could determine key parameters for assessing tidal energy dissipation, apsidal precession mechanisms, variations due to systemic proper motion, and the dynamical effects of non-resonant companion objects, depending on the applied models.
 
-`OrbDot` remains highly efficient with multiple data types and a high number of free parameters, as it utilizes the powerful nested sampling algorithms of the `Nestle` [@nestle; @Skilling:2006] and `PyMultiNest` [@Buchner:2014; @Feroz:2009] packages. The intricacies of the implementation are abstracted such that the `OrbDot` input files are simple and the method calls require only a list of free parameters, along with the desired model for fitting.
+`OrbDot` remains highly efficient with multiple data types and a high number of free parameters, as it utilizes powerful nested sampling algorithms of the `Nestle` [@nestle; @Skilling:2006] and `PyMultiNest` [@Buchner:2014; @Feroz:2009] packages. The intricacies of the implementation are abstracted such that `OrbDot` input files are simple and the method calls require only a list of free parameters, along with the desired model for fitting.
 
 Extensive documentation, including examples, is hosted on [ReadTheDocs](https://orbdot.readthedocs.io). 
 
@@ -49,7 +49,7 @@ Many exoplanet systems now have transit and radial velocity data spanning over a
 `OrbDot` lowers the barrier to entry for researchers at all levels, including undergraduates, by making advanced statistical methods accessible without requiring extensive computational experience. Despite its ease of use, `OrbDot` is not intended to be a black box. Rather, with extensive documentation, examples, and accessible source code, it is presented to the community with transparency that lends itself to community contributions and independent verification of results. It is designed to be easily extendable, as the nested sampling framework supports custom log-likelihood models with free parameters that are part of the `OrbDot` ecosystem. This ensures that the software may evolve to meet the needs of the research community.
 
 # Similar software
[…]
