+++ joss/paper.bib
@@ -53,6 +53,7 @@
          year = 2021,
         month = mar,
           eid = {ascl:2103.022},
+       url = {https://ascl.net/2103.022},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2021ascl.soft03022B},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
@@ -67,6 +68,7 @@
         month = oct,
           eid = {402.02},
         pages = {402.02},
+          doi = {10.3847/25c2cfeb.64cecc52},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2024DPS....5640202B},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
@@ -77,7 +79,7 @@
         year = {2025},
    publisher = {GitHub},
      journal = {GitHub repository},
-howpublished = {\url{https://github.com/BoiseStatePlanetary/susie.git}},
+howpublished = {\url{https://github.com/BoiseStatePlanetary/susie}},
 }
 
 @ARTICLE{Wang:2024,
@@ -162,6 +164,7 @@ archivePrefix = {arXiv},
          year = 2019,
         month = jun,
           eid = {ascl:1906.004},
+          url = {http://ascl.net/1906.004},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2019ascl.soft06004T},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
