### 2025-09-02 fe3ba75
JOSS paper typos
