+++ joss/paper.md
@@ -35,7 +35,7 @@ The necessity to search over many orbital epochs coupled with an ever-growing ar
 
 A key advantage of `OrbDot` is its ability to fit multiple types of data simultaneously, which can help to break parameter degeneracies. It also excels at assisting in result interpretation by generating reports on model comparisons and assessments of various physical effects in the context of the models and their corresponding theory. For example, analysis reports could determine key parameters for assessing tidal energy dissipation, apsidal precession mechanisms, variations due to systemic proper motion, and the dynamical effects of non-resonant companion objects, depending on the applied models.
 
-`OrbDot` remains highly efficient with multiple data types and a high number of free parameters, as it utilizes powerful nested sampling algorithms of the `Nestle` [@nestle; @Skilling:2006] and `PyMultiNest` [@Buchner:2014; @Feroz:2009] packages. The intricacies of the implementation are abstracted such that `OrbDot` input files are simple and the method calls require only a list of free parameters, along with the desired model for fitting.
+`OrbDot` remains highly efficient with multiple data types and a high number of free parameters, as it utilizes powerful nested sampling algorithms of the `nestle` [@nestle; @Skilling:2006] and `PyMultiNest` [@Buchner:2014; @Feroz:2009] packages. The intricacies of the implementation are abstracted such that the `OrbDot` input files are simple and the method calls require only a list of free parameters, along with the desired model for fitting.
 
 Extensive documentation, including examples, is hosted on [ReadTheDocs](https://orbdot.readthedocs.io). 
 
@@ -51,7 +51,7 @@ Many exoplanet systems now have transit and radial velocity data spanning over a
 # Similar software
 Some existing tools have features that overlap with `OrbDot`’s functionality, but none provides its full suite of capabilities. The most similar codes, `Susie` [@susie; @Barker:2024] and `PdotQuest` [@Wang:2024], which are not fully packaged software, are designed to fit secular evolution models to transit and eclipse timing data, but not to radial velocities. `OrbDot` also has greater flexibility in model fitting than `Susie` and `PdotQuest`. For example, `Susie` [@susie; @Barker:2024] employs simple least-squares fitting, and while `PdotQuest` [@Wang:2024] uses MCMC, it currently supports only the orbital decay model, making both codes narrower in scope than `OrbDot`. Moreover, `OrbDot` includes tools for theoretical interpretation.
 
-The well-known package TTVFast [@Deck:2014] is highly robust and capable of modeling both transit and radial velocity data, but it is focused on short-term timing variations driven by multi-planet dynamics near mean-motion resonances. `OrbDot`, in contrast, explores long-term secular models. Similarly, general-purpose frameworks such as `Juliet` [@Espinoza:2019], `exoplanet` [@ForemanMackey:2021], `ExoFast` [@Eastman:2019], `ExoStriker` [@Trifonov:2019], and `allesfitter` [@Gunther:2021] can jointly model transit and RV data, but their TTV models are also restricted to short-term variations, with an emphasis on transit light curve fitting rather than directly using transit and eclipse mid-times to constrain models.
[…]
