### 2025-09-02 07909ef
updated citation info and PyPi badge
