+++ docs/source/citing-orbdot.rst
@@ -4,12 +4,7 @@
 Citing OrbDot
 *************
 
-.. important::
-    The links below are not yet live... OrbDot code paper and TrES-1 b case study coming soon!
-
-OrbDot is free software made available under the MIT License. If you use OrbDot in your research or projects, please cite the `JOSS paper <LINK>`__  and the `GitHub repository <https://github.com/simonehagey/orbdot>`__ .
-
-If relevant to your work, please also consider citing the `complementary case study <LINK>`__ of Hot Jupiter TrES-1 b, which demonstrates the full capabilities of the OrbDot package.
+OrbDot is free software made available under the MIT License. If you use OrbDot in your research, please cite the `JOSS paper <https://joss.theoj.org/papers/10.21105/joss.08550>`__  and the `GitHub repository <https://github.com/simonehagey/orbdot>`__.
 
 BibTeX
 ------
@@ -18,19 +13,26 @@ You can cite OrbDot using the following BibTeX entries:
 .. code-block::
 
     @article{Hagey2025,
-      author       = {Simone R. Hagey, Aaron C. Boley},
-      title        = {``OrbDot``: A Python package for studying the secular evolution of exoplanet orbits},
-      journal      = {JOSS},
-      url          = {...},
-      year         = {2025},
+           author = {{Hagey}, Simone R. and {Boley}, Aaron},
+            title = "{``OrbDot``: A Python package for studying the secular evolution of exoplanet orbits}",
+          journal = {The Journal of Open Source Software},
+             year = 2025,
+            month = sep,
+           volume = {10},
+           number = {113},
+            pages = {8550},
+              doi = {10.21105/joss.08550},
     }
 
 .. code-block::
 
     @software{orbdot,
-      author       = {Simone R. Hagey},
-      title        = {A Python package for studying secular evolution of exoplanet orbits using observational data},
-      version      = {1.0.0},
-      url          = {https://github.com/simonehagey/orbdot},
-      year         = {2025},
+           author = {{Hagey}, Simone R. and {Boley}, Aaron},
+            title = {simonehagey/orbdot: v1.0.0},
+            month = sep,
+            year = 2025,
+        publisher = {Zenodo},
+          version = {v1.0.0},
+              doi = {10.5281/zenodo.17033860},
+              url = {https://doi.org/10.5281/zenodo.17033860},
     }
