### 2025-09-02 #8550 @simonehagey
Hi @warrickball, apologies for the delay in getting back to this, some unexpected things came up.

As suggested by @soichiro-hattori and @j-faria, I added a new section at the end of the JOSS paper discussing similar software. I hope it’s not too long!

I made a new release of OrbDot (v1.0.0), which is archived on Zenodo ([https://doi.org/10.5281/zenodo.17033860](https://doi.org/10.5281/zenodo.17033860?utm_source=chatgpt.com)) and available via pip ([https://pypi.org/project/orbdot/1.0.0/](https://pypi.org/project/orbdot/1.0.0/?utm_source=chatgpt.com)).

I also made some minor edits to the documentation to update links and citation information for the JOSS and AAS papers.

Let me know if there is anything else I need to do! Thanks, Simone.
