### 2025-09-02 edb7ace
updated links/bib entry for published TrES-1 b AAS paper
