+++ joss/paper.bib
@@ -214,9 +214,15 @@ archivePrefix = {arXiv},
 }
 
 @ARTICLE{Hagey:2025,
-       author = {{Hagey}, Simone R. and {Edwards}, Billy and {Tsiaras}, Angelos and and {Boley}, Aaron and others},
         title = "{TrES-1 b: A Case Study in Detecting Secular Evolution of Exoplanet Orbits}",
+       author = {Hagey, Simone R. and Edwards, Billy and Tsiaras, Angelos and Boley, Aaron C. and Kokori, Anastasia and Narita, Norio and Sada, Pedro V. and Walter, Filip and Zellem, Robert T. and A-thano, Napaporn and Alton, Kevin B. and Álava Amat, Miguel Ángel and Benni, Paul and Besson, Emmanuel and Brandebourg, Patrick and Bretton, Marc and Caló, Mauro and Crow, Martin Valentine and Dalouzy, Jean-Christophe and Deldem, Marc and Eenmäe, Tõnis and Ferratfiat, Stephane and Guerra, Pere and Haagen, Gary Vander and Hose, Ken and Jones, Adrian and Jongen, Yves and Laloum, Didier and Lora, Stefano and Marchini, Alessandro and Michelet, Jacques and Mihelčič, Matej and Mieglitz, Johannes and Miny, Eric and Molina, David and Morales Aimar, Mario and Nicollerat, Raphael and Peretto, Ivo and Raetz, Manfred and Regembal, François and Roth, Robert and Rousselot, Lionel and Salisbury, Mark and Sergison, Darryl and Wünsche, Anaël and Trnka, Jaroslav},
       journal = {The Astronomical Journal},
          year = 2025,
+        month = {sep},
+    publisher = {The American Astronomical Society},
+       volume = {170},
+       number = {4},
+        pages = {197},
           doi = {10.3847/1538-3881/aded15},
+          url = {https://dx.doi.org/10.3847/1538-3881/aded15},
 }
