+++ README.rst
@@ -56,6 +56,6 @@ Feedback, bug reports, feature requests, and code contributions are welcome! Ple
 
 TrES-1 b Case Study
 ===================
-In addition to the documentation, there is a complementary `TrES-1 b case study <https://arxiv.org/abs/2508.15075>`__ that showcases the full functionality of OrbDot’s model-fitting and interpretive tools, while providing a deeper dive into the theoretical foundations behind the package. This paper consolidates key equations in one place, making it a valuable reference for researchers looking to apply OrbDot to their own work.
+In addition to the documentation, there is a complementary `TrES-1 b case study <https://doi.org/10.3847/1538-3881/aded15>`__ that showcases the full functionality of OrbDot’s model-fitting and interpretive tools, while providing a deeper dive into the theoretical foundations behind the package. This paper consolidates key equations in one place, making it a valuable reference for researchers looking to apply OrbDot to their own work.
 
 
