### 2025-09-02 #8550 @warrickball
Could you update the title of the Zenodo archive to match the paper? (I.e., *OrbDot: A Python package for studying the secular evolution of exoplanet orbits*.) There should be an option to edit the metadata; you *don't* need to create a new archive.

I'm aiming to double check my editorial changes in the next hour or so and will open a pull request. Once those changes are in (or you make appropriate ones yourself), we'll give the production pipeline a test run.
