### 2025-09-02 #8550 @simonehagey
> Could you update the title of the Zenodo archive to match the paper? (I.e., _OrbDot: A Python package for studying the secular evolution of exoplanet orbits_.) There should be an option to edit the metadata; you _don't_ need to create a new archive.

@warrickball Done!

> I'm aiming to double check my editorial changes in the next hour or so and will open a pull request. Once those changes are in (or you make appropriate ones yourself), we'll give the production pipeline a test run.

Sounds good, thank you!
