### 2025-09-02 affd9f6
Initial editorial tweaks for JOSS paper
