+++ joss/paper.bib
@@ -1,6 +1,6 @@
 @ARTICLE{Feroz:2009,
        author = {{Feroz}, F. and {Hobson}, M.~P. and {Bridges}, M.},
-        title = "{MULTINEST: an efficient and robust Bayesian inference tool for cosmology and particle physics}",
+        title = {{MULTINEST}: an efficient and robust {Bayesian} inference tool for cosmology and particle physics},
       journal = {Monthly Notices of the Royal Astronomical Society},
          year = 2009,
         month = oct,
@@ -11,7 +11,7 @@
 }
 
 @ARTICLE{Skilling:2006,
-	    title = {Nested sampling for general {Bayesian} computation},
+        title = {Nested sampling for general {Bayesian} computation},
        volume = {1},
           doi = {10.1214/06-BA127},
        number = {4},
@@ -35,7 +35,7 @@
 
 @ARTICLE{Edwards:2023,
        author = {{Edwards}, Billy and {Changeat}, Quentin and {Tsiaras}, Angelos and {Allan}, Andrew and {Behr}, Patrick and {Hagey}, Simone R. and {Himes}, Michael D. and {Ma}, Sushuang and {Stassun}, Keivan G. and {Thomas}, Luis and {Thompson}, Alexandra and {Boley}, Aaron and {Booth}, Luke and {Bouwman}, Jeroen and {France}, Kevin and {Lowson}, Nataliea and {Meech}, Annabella and {Phillips}, Caprice L. and {Vidotto}, Aline A. and {Yip}, Kai Hou and {Bieger}, Michelle and {Gressier}, Am{\'e}lie and {Janin}, Estelle and {Jiang}, Ing-Guey and {Leonardi}, Pietro and {Sarkar}, Subhajit and {Skaf}, Nour and {Taylor}, Jake and {Yang}, Ming and {Ward-Thompson}, Derek},
-        title = "{Characterizing a World Within the Hot-Neptune Desert: Transit Observations of LTT 9779 b with the Hubble Space Telescope/WFC3}",
+        title = {Characterizing a World Within the Hot-{Neptune} Desert: Transit Observations of {LTT} 9779 b with the {Hubble Space Telescope/WFC3}},
       journal = {The Astronomical Journal},
          year = 2023,
         month = oct,
@@ -214,7 +214,7 @@ archivePrefix = {arXiv},
 }
 
 @ARTICLE{Hagey:2025,
-        title = "{TrES-1 b: A Case Study in Detecting Secular Evolution of Exoplanet Orbits}",
+        title = {{TrES-1 b}: A Case Study in Detecting Secular Evolution of Exoplanet Orbits},
[…]
