### 2025-09-02 #8550 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
