### 2025-09-02 #8550 @warrickball
Nearly there! I had a look at the paper of the weekend and as a result I missed the changes you made earlier today (specifically [these changes](https://github.com/simonehagey/orbdot/commit/08187264813031e3ef3768201b795d4f49130d77)). I'll review those in the next day or so and make another PR.
