### 2025-09-02 #8550 @simonehagey
> Nearly there! I had a look at the paper of the weekend and as a result I missed the changes you made earlier today (specifically [these changes](https://github.com/simonehagey/orbdot/commit/08187264813031e3ef3768201b795d4f49130d77)). I'll review those in the next day or so and make another PR.

@warrickball No worries, I’m sorry that I didn’t clarify the timing of those changes!

I read over the proof and it looks great so far. Do you think it is okay to submit to the arXiv, or should I wait for the final version? I’m not sure what is allowed. Thanks!
