### 2025-09-03 6713198
Further editorial tweaks for JOSS paper
