+++ joss/paper.bib
@@ -75,7 +75,7 @@
 
 @MISC{susie,
       author = {{Barker}, Malia and {Kirk}, Adrienne},
-       title = {The Susie Python Package},
+       title = {The {Susie} {Python} Package},
         year = {2025},
    publisher = {GitHub},
      journal = {GitHub repository},
@@ -84,7 +84,7 @@ howpublished = {\url{https://github.com/BoiseStatePlanetary/susie}},
 
 @ARTICLE{Wang:2024,
        author = {{Wang}, Wenqin and {Zhang}, Zixin and {Chen}, Zhangliang and {Wang}, Yonghao and {Yu}, Cong and {Ma}, Bo},
-        title = "{Long-term Variations in the Orbital Period of Hot Jupiters from Transit-timing Analysis Using TESS Survey Data}",
+        title = {Long-term Variations in the Orbital Period of Hot {Jupiters} from Transit-timing Analysis Using {TESS} Survey Data},
       journal = {The Astrophysical Journal Supplement Series},
          year = 2024,
         month = jan,
@@ -99,7 +99,7 @@ howpublished = {\url{https://github.com/BoiseStatePlanetary/susie}},
 
 @ARTICLE{Deck:2014,
        author = {{Deck}, Katherine M. and {Agol}, Eric and {Holman}, Matthew J. and {Nesvorn{\'y}}, David},
-        title = "{TTVFast: An Efficient and Accurate Code for Transit Timing Inversion Problems}",
+        title = {{TTVFast}: An Efficient and Accurate Code for Transit Timing Inversion Problems},
       journal = {The Astrophysical Journal},
          year = 2014,
         month = jun,
@@ -114,7 +114,7 @@ howpublished = {\url{https://github.com/BoiseStatePlanetary/susie}},
 
 @ARTICLE{Espinoza:2019,
        author = {{Espinoza}, N{\'e}stor and {Kossakowski}, Diana and {Brahm}, Rafael},
-        title = "{juliet: a versatile modelling tool for transiting and non-transiting exoplanetary systems}",
+        title = {{juliet}: a versatile modelling tool for transiting and non-transiting exoplanetary systems},
       journal = {Monthly Notices of the Royal Astronomical Society},
          year = 2019,
         month = dec,
@@ -128,8 +128,8 @@ howpublished = {\url{https://github.com/BoiseStatePlanetary/susie}},
 
 @ARTICLE{ForemanMackey:2021,
        author = {{Foreman-Mackey}, Daniel and {Luger}, Rodrigo and {Agol}, Eric and {Barclay}, Thomas and {Bouma}, Luke and {Brandt}, Timothy and {Czekala}, Ian and {David}, Trevor and {Dong}, Jiayin and {Gilbert}, Emily and {Gordon}, Tyler and {Hedges}, Christina and {Hey}, Daniel and {Morris}, Brett and {Price-Whelan}, Adrian and {Savel}, Arjun},
-        title = "{exoplanet: Gradient-based probabilistic inference for exoplanet data \& other astronomical time series}",
-      journal = {The Journal of Open Source Software},
+        title = {{exoplanet}: Gradient-based probabilistic inference for exoplanet data \& other astronomical time series},
+      journal = {Journal of Open Source Software},
          year = 2021,
         month = jun,
        volume = {6},
@@ -143,7 +143,7 @@ howpublished = {\url{https://github.com/BoiseStatePlanetary/susie}},
 
 @ARTICLE{Eastman:2019,
[…]
