### 2025-09-03 #8550 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
